requirements.txt
flask
openai
pandas
python-dotenv